FeedController -> POSTMAN 으로 TEST 완료

+++ src/main/java/com/rocket/user/userfeed/controller/FeedController.java
@@ -1,6 +1,7 @@
 package com.rocket.user.userfeed.controller;
 
 import static com.rocket.error.type.AuthErrorCode.INVALID_ACCESS_TOKEN;
+import static com.rocket.error.type.UserFeedErrorCode.FEED_COMMENT_USER_NOT_MATCH;
 
 import com.rocket.common.response.PageResponse;
 import com.rocket.config.jwt.AuthToken;
@@ -10,13 +11,13 @@ import com.rocket.error.exception.UserFeedException;
 import com.rocket.error.type.UserFeedErrorCode;
 import com.rocket.user.user.entity.User;
 import com.rocket.user.userfeed.dto.FeedCommentDto;
+import com.rocket.user.userfeed.dto.FeedDto;
 import com.rocket.user.userfeed.dto.FeedSearchCondition;
 import com.rocket.user.userfeed.entity.Feed;
 import com.rocket.user.userfeed.entity.FeedComment;
 import com.rocket.user.userfeed.entity.FeedCommentLike;
 import com.rocket.user.userfeed.service.FeedCommentLikeService;
 import com.rocket.user.userfeed.service.FeedCommentService;
-import com.rocket.user.userfeed.service.FeedImageService;
 import com.rocket.user.userfeed.service.FeedLikeService;
 import com.rocket.user.userfeed.service.FeedService;
 import com.rocket.user.userfeed.vo.FeedCommentResponse;
@@ -27,6 +28,7 @@ import com.rocket.utils.HeaderUtil;
 import io.jsonwebtoken.Claims;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -34,9 +36,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
@@ -58,28 +60,34 @@ public class FeedController {
 
     private final AuthTokenProvider tokenProvider;
 
-    private final FeedImageService feedImageService;
-
-
-    /**
-     * front에서 보내는 방법 참고: https://jaimemin.tistory.com/2072
-     */
     @PostMapping(consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
     public ApiResult<FeedResponse> createFeed(HttpServletRequest request
         , @RequestPart("files") List<MultipartFile> multipartFiles
         , @RequestPart("feed") Feed feed) {
+        User user = getUser(request);
 
         if (multipartFiles == null || multipartFiles.isEmpty()) {
             throw new UserFeedException(UserFeedErrorCode.UPLOAD_AT_LEAST_ONE_IMAGE);
         }
 
-        User user = getUser(request);
-        feedService.createFeed(user, feed, multipartFiles);
-        feedImageService.createFeedImage(user, feed, multipartFiles);
+        FeedDto newFeed = feedService.createFeed(user, feed, multipartFiles);
 
         return ApiUtils.success(FeedResponse.builder()
-            .feedId(String.valueOf(feed.getId()))
+            .feedId(String.valueOf(newFeed.getId()))
             .build());
+    }
+
+    @GetMapping("/{feedId}")
+    public ApiResult<FeedResponse> getFeed(HttpServletRequest request
+        , @PathVariable("feedId") String feedId) {
+        User user = getUser(request);
+        Feed feed = feedService.getFeed(Long.valueOf(feedId));
+
+        if (!feed.getUser().getId().equals(user.getId())) {
+            throw new UserFeedException(UserFeedErrorCode.FEED_USER_NOT_MATCH);
+        }
+
+        return ApiUtils.success(getFeedResponse(user, feed));
     }
 
     @GetMapping
@@ -128,58 +136,46 @@ public class FeedController {
             .build());
     }
 
-    @GetMapping("/{feedId}")
-    public ApiResult<FeedResponse> getFeed(HttpServletRequest request
-        , @PathVariable("feedId") String feedId) {
+    @PatchMapping(value = "/{feedId}")
+    public ApiResult<FeedResponse> updateFeed(HttpServletRequest request
+        , @PathVariable("feedId") String feedId
+        , @RequestPart("feed") FeedDto updateFeed) {
+
         User user = getUser(request);
         Feed feed = feedService.getFeed(Long.valueOf(feedId));
 
-        if (feed == null) {
-            throw new UserFeedException(UserFeedErrorCode.FEED_NOT_FOUND);
+        if (!feed.getUser().getId().equals(user.getId())) {
+            throw new UserFeedException(UserFeedErrorCode.FEED_USER_NOT_MATCH);
+        } else {
+            feedService.updateFeed(feed.getId(), updateFeed);
         }
-        return ApiUtils.success(getFeedResponse(user, feed));
-    }
 
-//    @PatchMapping(value = "/{feedId}", consumes = {MediaType.APPLICATION_JSON_VALUE,
-//        MediaType.MULTIPART_FORM_DATA_VALUE})
-//    public ApiResult<FeedResponse> updateImagePaths(HttpServletRequest request
-//        , @PathVariable("feedId") String feedId
-//        , @RequestPart("files") List<MultipartFile> multipartFiles) {
-//
-//        User user = getUser(request);
-//        FeedImage feedImage = feedImageService.getFeedImage(Long.valueOf(feedId));
-//        // TODO: feedImageService에서 S3 imagePath 수정하는 코드 구현 필요
-//
-//        // TODO: feedImageService에서 불러온 feedImage에서 feedId를 꺼내서 해당 id를 기반으로 Feed도 불러옴
-//        Feed feed = feedService.getFeed(Long.valueOf(feedId));
-//
-//        if (feed == null) {
-//            // TODO: UserFeedException 처리
-//            throw new UserFeedException(UserFeedErrorCode.FEED_NOT_FOUND);
-//        }
-//
-//        return ApiUtils.success(FeedResponse.builder()
-////             .feedId()
-////             .title()
-////             .content()
-////             .imagePaths()
-//            .build());
-//    }
+        return ApiUtils.success(FeedResponse.builder()
+            .feedId(String.valueOf(feed.getId()))
+            .title(feed.getTitle())
+            .content(feed.getContent())
+            .rcate1(feed.getRcate1())
+            .rcate2(feed.getRcate2())
+            .longitude(feed.getLongitude())
+            .latitude(feed.getLatitude())
+            .build());
+    }
 
     @DeleteMapping("/{feedId}")
     public ApiResult deleteFeed(HttpServletRequest request
         , @PathVariable("feedId") String feedId) {
         User user = getUser(request);
+
         feedService.deleteFeed(user.getId(), Long.valueOf(feedId));
 
         return ApiUtils.success(null);
     }
 
     @PostMapping("/{feedId}/like")
-    public ApiResult addFeedLike(HttpServletRequest request
+    public ApiResult saveFeedLike(HttpServletRequest request
         , @PathVariable("feedId") String feedId) {
         User user = getUser(request);
-        feedLikeService.createFeedLike(user, Long.valueOf(feedId));
+        feedLikeService.saveFeedLike(user, Long.valueOf(feedId));
 
         return ApiUtils.success(null);
     }
@@ -197,7 +193,8 @@ public class FeedController {
     public ApiResult<FeedCommentResponse> createFeedComment(
         HttpServletRequest request
         , @PathVariable("feedId") String feedId
-        , @RequestBody FeedCommentDto feedCommentDto) {
+        , @RequestPart("feedComment") FeedCommentDto feedCommentDto) {
+
         User user = getUser(request);
         Feed feed = feedService.getFeed(Long.valueOf(feedId));
 
@@ -214,7 +211,7 @@ public class FeedController {
     }
 
     @GetMapping("/{feedId}/comments")
-    public ApiResult<PageResponse<FeedCommentResponse>> getFeedComments(HttpServletRequest request
+    public ApiResult<PageResponse<FeedResponse>> getFeedComments(HttpServletRequest request
         , @PathVariable("feedId") String feedId
         , Pageable pageable) {
         User user = getUser(request);
@@ -227,21 +224,21 @@ public class FeedController {
         Page<FeedComment> feedComments = feedCommentService.getFeedComments(
             Long.valueOf(feedId), pageable);
 
-        List<FeedCommentResponse> content = new ArrayList<>();
+        List<FeedResponse> content = new ArrayList<>();
 
-//        for (FeedComment feedComment : feedComments) {
-//            content.add(FeedResponse.builder()
-//                .userId(String.valueOf(user.getId()))
-//                .profileImagePath(user.getProfileImage())
-//                .nickname(user.getNickname())
-//                .email(user.getEmail())
-//                .content(feedComment.getComment())
-//                .commentLikeCnt(
-//                    feedCommentLikeService.getCount(feedComment.getId()))
-//                .build());
-//        }
+        for (FeedComment feedComment : feedComments) {
+            content.add(FeedResponse.builder()
+                .userId(String.valueOf(user.getId()))
+                .profileImagePath(user.getProfileImage())
+                .nickname(user.getNickname())
+                .email(user.getEmail())
+                .content(feedComment.getComment())
+                .commentLikeCnt((long) feedComment.getFeedCommentLike().size())
 
-        return ApiUtils.success(PageResponse.<FeedCommentResponse>builder()
+                .build());
+        }
+
+        return ApiUtils.success(PageResponse.<FeedResponse>builder()
             .lastPage(feedComments.isLast())
             .firstPage(feedComments.isFirst())
             .totalPages(feedComments.getTotalPages())
@@ -252,14 +249,42 @@ public class FeedController {
             .build());
     }
 
+    @PatchMapping(value = "/{feedId}/comments/{commentId}")
+    public ApiResult<FeedCommentResponse> updateFeedComment(HttpServletRequest request
+        , @PathVariable("feedId") String feedId
+        , @PathVariable("commentId") String commentId
+        , @RequestPart("feedComment") FeedCommentDto updateFeedComment) {
+
+        User user = getUser(request);
+        FeedComment feedComment = feedCommentService.getFeedComment(Long.valueOf(commentId));
+
+        if (!feedComment.getUser().getId().equals(user.getId())) {
+            throw new UserFeedException(UserFeedErrorCode.FEED_USER_NOT_MATCH);
+        } else {
+            feedCommentService.updateFeedComment(Long.valueOf(commentId), updateFeedComment);
+        }
+
+        return ApiUtils.success(FeedCommentResponse.builder()
+            .comment(updateFeedComment.getComment())
+            .build());
+    }
+
     @DeleteMapping("/{feedId}/comments/{commentId}")
-    public ApiResult<FeedCommentDto> deleteComment(@PathVariable("feedId") String feedId
+    public ApiResult<FeedCommentDto> deleteFeedComment(HttpServletRequest request
+        , @PathVariable("feedId") String feedId
         , @PathVariable("commentId") String commentId) {
+
+        User user = getUser(request);
+
         FeedComment feedComment
             = feedCommentService.getFeedComment(Long.valueOf(commentId));
+
+        if (!user.getId().equals(feedComment.getUser().getId())) {
+            throw new UserFeedException(FEED_COMMENT_USER_NOT_MATCH);
+        }
+
         feedCommentService.deleteFeedComment(feedComment);
 
-        // TODO: 실패 케이스에 대해서 구현 필요
         return ApiUtils.success(FeedCommentDto.builder()
             .feedCommentId(feedComment.getId())
             .comment(feedComment.getComment())
@@ -267,11 +292,12 @@ public class FeedController {
     }
 
     @PostMapping("/{feedId}/comments/{commentId}/like")
-    public ApiResult likeComment(HttpServletRequest request
+    public ApiResult saveFeedCommentLike(HttpServletRequest request
         , @PathVariable("feedId") String feedId
         , @PathVariable("commentId") String commentId) {
         User user = getUser(request);
         FeedComment feedComment = feedCommentService.getFeedComment(Long.valueOf(commentId));
+
         feedCommentLikeService.saveFeedCommentLike(user, feedComment);
 
         // TODO: 실패 케이스에 대해서 구현 필요
@@ -279,7 +305,7 @@ public class FeedController {
     }
 
     @DeleteMapping("/{feedId}/comments/{commentId}/like")
-    public ApiResult unlikeComment(HttpServletRequest request
+    public ApiResult deleteFeedCommentLike(HttpServletRequest request
         , @PathVariable("feedId") String feedId
         , @PathVariable("commentId") String commentId) {
         User user = getUser(request);
@@ -311,12 +337,16 @@ public class FeedController {
             .longitude(feed.getLongitude())
             .latitude(feed.getLatitude())
 
-            .feedLikeCnt(feedLikeService.getCount(feed.getId())) // TODO: null 예외 처리 필요
-            .feedCommentCnt(feedCommentService.getCount(feed.getId()))
-            .feedImages(feedImageService.getFeedImages(feed.getId()))
+            .feedLikeCnt((long) feed.getFeedLike().size())
+            .feedCommentCnt((long) feed.getFeedComment().size())
+            .feedImages(feed.getFeedImage().stream()
+                .map(feedImage -> feedImage.getImagePaths()).collect(Collectors.toList()))
+            .build();
+//            .feedLikeCnt(feedLikeService.getCount(feed.getId())) // TODO: null 예외 처리 필요
+//            .feedCommentCnt(feedCommentService.getCount(feed.getId()))
+//            .feedImages(feedImageService.getFeedImages(feed.getId()))
 //            .comments(feedCommentService.getFeedComments(feed.getId(), Pageable.ofSize(10)))
 //            .commentLikeCnt(feedCommentLikeService.getCount(feed.getId()))
-            .build();
     }
 
     private FeedResponse getFeedListResponse(Feed feed) {
@@ -328,14 +358,12 @@ public class FeedController {
             .longitude(feed.getLongitude())
             .latitude(feed.getLatitude())
 
-            .feedLikeCnt(feedLikeService.getCount(feed.getId())) // TODO: null 예외 처리 필요
-            .feedCommentCnt(feedCommentService.getCount(feed.getId()))
-            .feedImages(feedImageService.getFeedImages(feed.getId()))
-//            .comments(feedCommentService.getFeedComments(feed.getId(), Pageable.ofSize(10)))
-//            .commentLikeCnt(feedCommentLikeService.getCount(feed.getId()))
+            .feedLikeCnt((long) feed.getFeedLike().size())
+            .feedCommentCnt((long) feed.getFeedComment().size())
+            .feedImages(feed.getFeedImage().stream()
+                .map(feedImage -> feedImage.getImagePaths()).collect(Collectors.toList()))
             .build();
     }
-
 
     private User getUser(HttpServletRequest request) {
         String accessToken = HeaderUtil.getAccessToken(request);
